Generated by Spark: i don't think the homepage needs to poll and update in real time. i just want that when a user comes to the homepage, it pulls in accurate stats for total games logged, investigators played, unique campaigns, popular campaigns, most played investigators, popular standalone scenarios, and popular side stories.
I haven't audited all of the data, but it seems like the popular side stories field did not update after i edited an entry for my user to add in a second play of "fortune and folly". can you please check the logic here. attached is a screenshot showing the community stats card for this showing only 1 play of fortune and folly, but the data should show 2 plays if i am correct.